chatbot-system/requirements.txt
flask
python-dotenv
sentence-transformers
torch
neo4j
openai